# integrations/openapi.json
{
  "components": {
    "schemas": {
      "ChatTurn": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ChatTurn",
        "type": "object"
      },
      "Diagnostic": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "default": "syntax",
            "title": "Code",
            "type": "string"
          },
          "column": {
            "title": "Column",
            "type": "integer"
          },
          "line": {
            "title": "Line",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "severity": {
            "default": "error",
            "enum": [
              "error",
              "warning"
            ],
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "line",
          "column",
          "message"
        ],
        "title": "Diagnostic",
        "type": "object"
      },
      "ExecutionContext": {
        "additionalProperties": false,
        "properties": {
          "algorithm": {
            "default": "",
            "maxLength": 120,
            "title": "Algorithm",
            "type": "string"
          },
          "arrays": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "array"
            },
            "maxProperties": 10,
            "title": "Arrays",
            "type": "object"
          },
          "code": {
            "default": "",
            "maxLength": 16000,
            "title": "Code",
            "type": "string"
          },
          "current_line": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Line"
          },
          "language_version": {
            "default": "avp-repository-v1",
            "maxLength": 80,
            "title": "Language Version",
            "type": "string"
          },
          "phase": {
            "default": "unknown",
            "enum": [
              "before",
              "after",
              "unknown"
            ],
            "title": "Phase",
            "type": "string"
          },
          "recent_events": {
            "items": {
              "$ref": "#/components/schemas/ExecutionEvent"
            },
            "maxItems": 8,
            "title": "Recent Events",
            "type": "array"
          },
          "run_id": {
            "default": "",
            "maxLength": 120,
            "title": "Run Id",
            "type": "string"
          },
          "step": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step"
          },
          "variables": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "maxProperties": 80,
            "title": "Variables",
            "type": "object"
          }
        },
        "title": "ExecutionContext",
        "type": "object"
      },
      "ExecutionEvent": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "maxLength": 500,
            "title": "Description",
            "type": "string"
          },
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "maxProperties": 20,
            "title": "Details",
            "type": "object"
          },
          "indices": {
            "items": {
              "type": "integer"
            },
            "maxItems": 20,
            "title": "Indices",
            "type": "array"
          },
          "kind": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "line": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          }
        },
        "required": [
          "kind",
          "description"
        ],
        "title": "ExecutionEvent",
        "type": "object"
      },
      "FeedbackRequest": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "default": "",
            "maxLength": 1000,
            "title": "Comment",
            "type": "string"
          },
          "interaction_id": {
            "maxLength": 100,
            "minLength": 20,
            "title": "Interaction Id",
            "type": "string"
          },
          "rating": {
            "enum": [
              "helpful",
              "incorrect",
              "too_much_answer",
              "too_vague",
              "missing_context"
            ],
            "title": "Rating",
            "type": "string"
          }
        },
        "required": [
          "interaction_id",
          "rating"
        ],
        "title": "FeedbackRequest",
        "type": "object"
      },
      "GenerateRequest": {
        "properties": {
          "query": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "GenerateRequest",
        "type": "object"
      },
      "GenerateResponse": {
        "properties": {
          "cached": {
            "default": false,
            "title": "Cached",
            "type": "boolean"
          },
          "generated_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated Code"
          },
          "retrieved_functions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RetrievedFunction"
            },
            "title": "Retrieved Functions",
            "type": "array"
          },
          "validation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidationResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "GenerateResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "knowledge_version": {
            "default": "",
            "title": "Knowledge Version",
            "type": "string"
          },
          "model": {
            "default": "",
            "title": "Model",
            "type": "string"
          },
          "note": {
            "default": "Configuration status only; model reachability is checked on request.",
            "title": "Note",
            "type": "string"
          },
          "provider": {
            "default": "",
            "title": "Provider",
            "type": "string"
          },
          "provider_configured": {
            "default": false,
            "title": "Provider Configured",
            "type": "boolean"
          },
          "retriever_loaded": {
            "default": false,
            "title": "Retriever Loaded",
            "type": "boolean"
          },
          "status": {
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "HealthResponse",
        "type": "object"
      },
      "JsonValue": {},
      "RetrieveRequest": {
        "properties": {
          "k": {
            "default": 2,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "K",
            "type": "integer"
          },
          "query": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "RetrieveRequest",
        "type": "object"
      },
      "RetrieveResponse": {
        "properties": {
          "cached": {
            "default": false,
            "title": "Cached",
            "type": "boolean"
          },
          "results": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RetrievedFunction"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "title": "RetrieveResponse",
        "type": "object"
      },
      "RetrievedFunction": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "function_name": {
            "title": "Function Name",
            "type": "string"
          },
          "parameters": {
            "items": {
              "type": "string"
            },
            "title": "Parameters",
            "type": "array"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "source_file": {
            "default": "",
            "title": "Source File",
            "type": "string"
          }
        },
        "required": [
          "score",
          "function_name",
          "parameters",
          "code"
        ],
        "title": "RetrievedFunction",
        "type": "object"
      },
      "Source": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "section": {
            "title": "Section",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "path",
          "section",
          "text"
        ],
        "title": "Source",
        "type": "object"
      },
      "TutorRequest": {
        "additionalProperties": false,
        "properties": {
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "feedback_consent": {
            "default": false,
            "title": "Feedback Consent",
            "type": "boolean"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ChatTurn"
            },
            "maxItems": 12,
            "title": "History",
            "type": "array"
          },
          "level": {
            "default": "beginner",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "title": "Level",
            "type": "string"
          },
          "mode": {
            "default": "explain",
            "enum": [
              "explain",
              "hint",
              "predict",
              "debug"
            ],
            "title": "Mode",
            "type": "string"
          },
          "question": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "TutorRequest",
        "type": "object"
      },
      "TutorResponse": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "context_status": {
            "enum": [
              "provided",
              "partial",
              "missing"
            ],
            "title": "Context Status",
            "type": "string"
          },
          "fallback_used": {
            "default": false,
            "title": "Fallback Used",
            "type": "boolean"
          },
          "harness_version": {
            "default": "",
            "title": "Harness Version",
            "type": "string"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "interaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interaction Id"
          },
          "knowledge_version": {
            "title": "Knowledge Version",
            "type": "string"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "skill": {
            "default": "",
            "title": "Skill",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "title": "Sources",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "answer",
          "context_status",
          "provider",
          "model",
          "latency_ms",
          "prompt_version",
          "knowledge_version"
        ],
        "title": "TutorResponse",
        "type": "object"
      },
      "UsageRequest": {
        "additionalProperties": false,
        "properties": {
          "algorithm": {
            "default": "",
            "maxLength": 120,
            "title": "Algorithm",
            "type": "string"
          },
          "consent": {
            "const": true,
            "title": "Consent",
            "type": "boolean"
          },
          "event": {
            "enum": [
              "offered",
              "opened",
              "dismissed",
              "asked",
              "prediction_correct",
              "prediction_incorrect"
            ],
            "title": "Event",
            "type": "string"
          }
        },
        "required": [
          "consent",
          "event"
        ],
        "title": "UsageRequest",
        "type": "object"
      },
      "ValidateRequest": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 16000,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "ValidateRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationResult": {
        "additionalProperties": false,
        "properties": {
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            },
            "title": "Diagnostics",
            "type": "array"
          },
          "execution_verified": {
            "default": false,
            "title": "Execution Verified",
            "type": "boolean"
          },
          "syntax_valid": {
            "title": "Syntax Valid",
            "type": "boolean"
          }
        },
        "required": [
          "syntax_valid"
        ],
        "title": "ValidationResult",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AVP Algorithm Tutor API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/context/validate": {
      "post": {
        "operationId": "validate_context_api_context_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionContext"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionContext"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Context"
      }
    },
    "/api/demo/insertion-sort": {
      "get": {
        "operationId": "demo_api_demo_insertion_sort_get",
        "parameters": [
          {
            "in": "query",
            "name": "values",
            "required": false,
            "schema": {
              "default": "2,7,9,4",
              "title": "Values",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Demo"
      }
    },
    "/api/feedback": {
      "post": {
        "operationId": "feedback_api_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feedback"
      }
    },
    "/api/feedback/{interaction_id}": {
      "delete": {
        "operationId": "delete_feedback_api_feedback__interaction_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "interaction_id",
            "required": true,
            "schema": {
              "title": "Interaction Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Feedback"
      }
    },
    "/api/generate": {
      "post": {
        "operationId": "generate_api_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate"
      }
    },
    "/api/health": {
      "get": {
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/api/retrieve": {
      "post": {
        "operationId": "retrieve_api_retrieve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrieveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retrieve"
      }
    },
    "/api/tutor": {
      "post": {
        "operationId": "tutor_api_tutor_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutorResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tutor"
      }
    },
    "/api/usage": {
      "post": {
        "operationId": "usage_api_usage_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Usage"
      }
    },
    "/api/validate": {
      "post": {
        "operationId": "validate_api_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate"
      }
    }
  }
}